Fixed all warnings generated when building the tracker project with the
	* src/libtracker-common/tracker-configuration.c: (string_replace):
	* src/libtracker-gtk/tracker-keyword-store.c:
	(tracker_keyword_store_drag_data_get):
	* src/libtracker-gtk/tracker-metadata-tile.c: (_int_to_label):
	* src/qdbm/myconf.c: (_qdbm_mmap):
	* src/text-filters/ooo_converter/ChangeLog:
	* src/text-filters/ooo_converter/o3read.c:
	* src/tracker-applet/tracker-applet.c: (reindex):
	* src/tracker-search-tool/tracker-search-tool.c:
	* src/trackerd/tracker-email-thunderbird.c:
	* src/trackerd/tracker-indexer.c:
	* src/trackerd/trackerd.c: (get_update_count): Fixed all warnings
	generated when building the tracker project with the exception of
	xdgmime which is imported. This fixes bug #520789.

svn path=/trunk/; revision=1225

+++ src/qdbm/myconf.c
@@ -250,7 +250,7 @@ void *_qdbm_mmap(void *start, size_t length, int prot, int flags, int fd, off_t 
   while((rv = read(fd, wp + rlen, length - rlen)) > 0){
     rlen += rv;
   }
-  if(rv == -1 || rlen != length){
+  if(rv == -1 || rlen != (int) length){
     free(buf);
     return MAP_FAILED;
   }
